FREEI-1975 Require composer libraries only once
Previously, composer libraries were included multiple times by the
certman library because we did not put the full path to the library
which would default to whatever module is being installed for some
reason. This commit provides the full path for the composer libraries
in reguards to the certman module. This way it should always include
the correct autoload class when loading the certman library